refactor(server): align config service with shared ConfigServiceImpl
- Extract server-config-store.ts: ConfigStore adapter with file I/O, env overrides, validation
- Replace ServerConfigService class with thin singleton wrapper around createConfigService
- Update bootstrap.ts to use loadServerConfig() + serverConfigService singleton
- Update media-processing-service.ts to use IConfigService interface
- Update server-config-service.test.ts for singleton pattern

PR: #317
